requirements.txt
Django>=1.11,<2.0
django-cors-headers>=2.1
django-debug-toolbar>=1.8
django-filter>=1.1.0
django-rest-swagger>=2.1.0
djangorestframework>=3.6.4
Markdown>=2.6.7
celery>=4.1.0
orangengine
